# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend escalable en Go para control financiero y automatización de reportes a WhatsApp.",
        "title": "Braille Web - API de Transacciones y Reportes",
        "contact": {},
        "version": "1.0"
    },
    "host": "api.example.com",
    "basePath": "/",
    "paths": {
        "/api/export": {
            "get": {
                "description": "Sincroniza y descarga una sábana de datos procesada en formato binario legible por Microsoft Excel o plataformas externas de contabilidad.",
                "produces": [
                    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet"
                ],
                "tags": [
                    "export"
                ],
                "summary": "Exportar historial de transacciones a Excel (.xlsx)",
                "responses": {
                    "200": {
                        "description": "Descarga de archivo reporte_brailer.xlsx",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "500": {
                        "description": "Error al construir dinámicamente la estructura Excel",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/summary": {
            "get": {
                "description": "Calcula el balance acumulado general, los gastos del mes corriente en base a la fecha actual y la cantidad de transacciones registradas.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "metrics"
                ],
                "summary": "Obtener resumen analítico",
                "responses": {
                    "200": {
                        "description": "Ejemplo: {'total_general': 5000.5, 'total_mes': 1200.0, 'conteo': 14}",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Error de servidor al calcular la lista",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/whatsapp/send": {
            "post": {
                "description": "Encola de forma asíncrona y segura el envío estructurado de un reporte hacia un canal o grupo de WhatsApp parametrizado. Evita duplicados simultáneos.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "whatsapp"
                ],
                "summary": "Reportar transacción a WhatsApp",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID único de la transacción a reportar",
                        "name": "id",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Nombre o JID del destinatario (Por defecto: 'Prueba')",
                        "name": "to",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estado de la transacción en cola: 'queued' o 'processing'",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Transacción o Grupo objetivo no encontrado",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "503": {
                        "description": "El cliente de WhatsApp no se encuentra inicializado o conectado",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
...
}